refactor(docx-reader): move part lookups into rels/parts.rs
